requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-jwt==1.19.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.3
pytest-django==4.9.0
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
